--- a/benchmark/compare-decode.js
+++ b/benchmark/compare-decode.js
@@ -1,86 +1,82 @@
 import fs from 'fs'
 import path from 'path'
-import bench from 'nanobench'
+import { fileURLToPath } from 'url'
 
 import bencode from '../index.js'
+
 import bencoding from 'bencoding'
 import bncode from 'bncode'
 import btparse from 'btparse'
-import dht from 'dht.js/lib/dht/bencode'
+import dht from 'dht.js/lib/dht/bencode.js'
 import dhtBencode from 'dht-bencode'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 const buffer = fs.readFileSync(path.join(__dirname, 'test.torrent'))
 
-const ITERATIONS = 10000
+const ITERATIONS = 1
 
-bench(`bencode.decode() ⨉ ${ITERATIONS}`, function (run) {
+suite('compare decode', () => {
+
+scenario(`bencode.decode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bencode.decode(buffer)
   }
-  run.end()
 
   return result
 })
 
-bench(`bencoding.decode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`bencoding.decode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bencoding.decode(buffer)
   }
-  run.end()
 
   return result
 })
 
-bench(`bncode.decode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`bncode.decode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bncode.decode(buffer)
   }
-  run.end()
 
   return result
 })
 
-bench(`btparse() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`btparse()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = btparse(buffer)
   }
-  run.end()
 
   return result
 })
 
-bench(`dht.decode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`dht.decode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = dht.decode(buffer)
   }
-  run.end()
 
   return result
 })
 
-bench(`dhtBencode.decode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`dhtBencode.decode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = dhtBencode.bdecode(buffer)
   }
-  run.end()
 
   return result
+})
+
 })