--- a/benchmark/compare-encode.js
+++ b/benchmark/compare-encode.js
@@ -1,74 +1,72 @@
 import fs from 'fs'
 import path from 'path'
-import bench from 'nanobench'
+import { fileURLToPath } from 'url'
 
 import bencode from '../index.js'
+
 import bencoding from 'bencoding'
 import bncode from 'bncode'
-import dht from 'dht.js/lib/dht/bencode'
+import dht from 'dht.js/lib/dht/bencode.js'
 import dhtBencode from 'dht-bencode'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 const buffer = fs.readFileSync(path.join(__dirname, 'test.torrent'))
 const object = bencode.decode(buffer)
 
-const ITERATIONS = 10000
+const ITERATIONS = 1
 
-bench(`bencode.encode() ⨉ ${ITERATIONS}`, function (run) {
+suite('compare encode', () => {
+
+scenario(`bencode.encode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bencode.encode(object)
   }
-  run.end()
 
   return result
 })
 
-bench(`bencoding.encode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`bencoding.encode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bencoding.encode(object)
   }
-  run.end()
 
   return result
 })
 
-bench(`bncode.encode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`bncode.encode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = bncode.encode(object)
   }
-  run.end()
 
   return result
 })
 
-bench(`dht.encode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`dht.encode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = dht.encode(object)
   }
-  run.end()
 
   return result
 })
 
-bench(`dhtBencode.encode() ⨉ ${ITERATIONS}`, function (run) {
+scenario(`dhtBencode.encode()`, function () {
   let result = null
 
-  run.start()
   for (let i = 0; i < ITERATIONS; i++) {
     result = dhtBencode.bencode(object)
   }
-  run.end()
 
   return result
+})
+
 })