--- a/benchmark/reference-after.js
+++ b/benchmark/reference-after.js
@@ -1,0 +1,3 @@
+import { fibonacciSuite } from './fibonacci.js'
+
+fibonacciSuite('reference after benchmarks')