--- a/benchmark/reference-before.js
+++ b/benchmark/reference-before.js
@@ -1,0 +1,4 @@
+import { fibonacciSuite } from './fibonacci.js'
+
+fibonacciSuite('reference before benchmarks')
+