--- a/package.json
+++ b/package.json
@@ -21,12 +21,12 @@
   "devDependencies": {
     "@webtorrent/semantic-release-config": "1.0.10",
     "bencoding": "latest",
+    "bipbip": "github:milahu/bipbip",
     "bncode": "latest",
     "browserify": "^17.0.0",
     "btparse": "latest",
     "dht-bencode": "latest",
     "dht.js": "latest",
-    "nanobench": "3.0.0",
     "semantic-release": "21.1.1",
     "standard": "17.1.0",
     "tap-spec": "5.0.0",
@@ -51,7 +51,8 @@
     "url": "git://github.com/webtorrent/node-bencode.git"
   },
   "scripts": {
-    "benchmark": "nanobench benchmark/*.js",
+    "benchmark": "bipbip --compare benchmark/benchmarks.json --save benchmark/benchmarks.json benchmark/benchmarks.js",
+    "benchmark-nosave": "bipbip --compare benchmark/benchmarks.json benchmark/benchmarks.js",
     "bundle": "mkdir -p dist && npm run bundle:lib && npm run bundle:test",
     "bundle:lib": "browserify lib/index.js -s bencode -o dist/bencode.js",
     "bundle:test": "browserify test/*.test.js -o dist/tests.js",